Switch to InventoryLoader from  Dialog subclass
